requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
sqlalchemy
psycopg2-binary
python-jose